AOC.Challenge

Description

Gather together all challenges and collect them into a single map.

Documentation

module AOC.Challenge.Day25

module AOC.Challenge.Day24

module AOC.Challenge.Day23

module AOC.Challenge.Day22

module AOC.Challenge.Day21

module AOC.Challenge.Day20

module AOC.Challenge.Day19

module AOC.Challenge.Day18

module AOC.Challenge.Day17

module AOC.Challenge.Day16

module AOC.Challenge.Day15

module AOC.Challenge.Day14

module AOC.Challenge.Day13

module AOC.Challenge.Day12

module AOC.Challenge.Day11

module AOC.Challenge.Day10

module AOC.Challenge.Day09

module AOC.Challenge.Day08

module AOC.Challenge.Day07

module AOC.Challenge.Day06

module AOC.Challenge.Day05

module AOC.Challenge.Day04

module AOC.Challenge.Day03

module AOC.Challenge.Day02

module AOC.Challenge.Day01

type ChallengeMap = Map (Finite 25) (Map Part SomeSolution)

A map of days to parts to solutions.

data ChallengeSpec

A specification for a specific challenge. Should consist of a day and a lowercase character.

Constructors

CS

Fields

_csDay :: Finite 25

_csPart :: Part

data Part

A given part of a problem. All Advent of Code challenges are two-parts.

You can usually get Part1 (if it is already released) with a nonsense session key, but Part2 always requires a valid session key.

Note also that Challenge #25 typically only has a single part.

Constructors

Part1

Part2

challengeMap :: ChallengeMap

A map of all challenges.

lookupSolution :: ChallengeSpec -> Map (Finite 25) (Map Part a) -> Maybe a

Lookup up a solution from a ChallengeMap

dayToInt :: Finite 25 -> Int

Turn a day represented by a Finite 25 into an integer day (1 - 25).

solSpec :: Name -> ChallengeSpec

Get a ChallengeSpec from a given reified solution (name).

solSpec 'day02a == CS { _csDay = 1, _csPart = a }

charPart :: Char -> Maybe Part

Parse a Literal into a Part