AOC.Challenge.Day02

Description

Day 2. See AOC.Solver for the types used in this module!

Documentation

day02a :: [String] :~> Int

We compute a frequency map of all of the characters in a string, and then get all of the frequencies that happened for each line.

Then we build a frequency map of the frequencies!

day02b :: [String] :~> String

The main work is in firstNeighbor, which looks thorugh a list of items and finds the first item whose neighbor was already seen.

Then we take the two "almost matching" strings and filter out all of the characters that aren't the same, using zipWith and catMaybes.