AOC.Challenge.Day13

Description

Day 13. See AOC.Solver for the types used in this module!

After completing the challenge, it is recommended to:

Replace AOC.Prelude imports to specific modules (with explicit imports) for readability.

Remove the -Wno-unused-imports and -Wno-unused-top-binds pragmas.

Replace the partial type signatures underscores in the solution types _ :~> _ with the actual types of inputs and outputs of the solution. You can delete the type signatures completely and GHC will recommend what should go in place of the underscores.

Documentation

type Point = V2 Int

data Track

Constructors

THoriz

TVert

TTurnNW

TTurnNE

TInter

type World = Map Point Track

data Dir

Constructors

DN

DE

DS

DW

data Cart

Constructors

C

Fields

_cPos :: Point

_cDir :: Dir

_cLast :: Maybe Point

_cTurns :: Int

parseWorld :: String -> (Set Cart, World)

(!?!) :: (Ord k, Show k) => Map k a -> k -> a

stepCart :: World -> Cart -> Cart

dirPoint :: Dir -> Point

turnWith :: Int -> Dir -> Dir

turnLeft :: Dir -> Dir

turnRight :: Dir -> Dir

stepCarts :: World -> Set Cart -> Either Point (Set Cart)

findCollision :: Set Cart -> Maybe Point

runWorld :: World -> Set Cart -> Maybe Point

day13a :: (World, Set Cart) :~> Point

stepCarts2 :: World -> Set Cart -> Either Point (Set Cart)

runWorld2 :: World -> Set Cart -> Maybe Point

day13b :: (World, Set Cart) :~> Point