AOC.Challenge.Day15

Description

Day 15. See AOC.Solver for the types used in this module!

After completing the challenge, it is recommended to:

Replace AOC.Prelude imports to specific modules (with explicit imports) for readability.

Remove the -Wno-unused-imports and -Wno-unused-top-binds pragmas.

Replace the partial type signatures underscores in the solution types _ :~> _ with the actual types of inputs and outputs of the solution. You can delete the type signatures completely and GHC will recommend what should go in place of the underscores.

Documentation

data EType

Constructors

EGob

EElf

data Entity

Constructors

E

Fields

_eType :: EType

_eHP :: Int

_eAtk :: Int

eType :: Lens' Entity EType

eHP :: Lens' Entity Int

eAtk :: Lens' Entity Int

type Point = V2 Int

type World = Set ScanPoint

type Entities = Map ScanPoint Entity

newtype ScanPoint

It's Point, but with a newtype wrapper so we have an Ord that sorts by y first, then x

Constructors

SP

Fields

_gSP :: Point

dist :: ScanPoint -> ScanPoint -> Int

neighbs :: ScanPoint -> Set ScanPoint

inRangeOf :: Set ScanPoint -> Set ScanPoint

type Path = [ScanPoint]

actualLiteralAStar :: World -> ScanPoint -> ScanPoint -> Maybe Path

Yes, it's actual literal A*.

stepTo

Arguments

:: Set ScanPoint
legal points

-> ScanPoint

-> ScanPoint

-> ScanPoint

isReachable :: Set ScanPoint -> ScanPoint -> ScanPoint -> Bool

stepEntity :: World -> Entities -> ScanPoint -> Entity -> Maybe ScanPoint

makeAttack :: World -> Entities -> ScanPoint -> Entity -> Maybe ScanPoint

data BattleLog a

Constructors

BLTurn (Maybe EType) a

BLRound a

BLOver Int

damage :: Int -> Maybe Entity -> (Maybe (First EType), Maybe Entity)

stepBattle :: World -> (Entities, Entities) -> BattleLog (Entities, Entities)

getOutcome :: BattleLog (Int, Int) -> (Int, Int)

day15a :: (World, Entities) :~> Int

elfKilled :: BattleLog Bool -> Bool

day15b :: (World, Entities) :~> Int

parseWorld :: String -> (World, Entities)

displayWorld :: World -> Entities -> String

boundingBox :: [Point] -> V2 Point