AOC.Challenge.Day23

Description

Day 23. See AOC.Solver for the types used in this module!

After completing the challenge, it is recommended to:

Replace AOC.Prelude imports to specific modules (with explicit imports) for readability.

Remove the -Wno-unused-imports and -Wno-unused-top-binds pragmas.

Replace the partial type signatures underscores in the solution types _ :~> _ with the actual types of inputs and outputs of the solution. You can delete the type signatures completely and GHC will recommend what should go in place of the underscores.

Documentation

type Point3 = V3 Int

data Circle

Constructors

C

Fields

_cCenter :: !Point3

_cRadius :: !Int

day23a :: _ :~> _

touchesRegion :: Circle -> V2 Point3 -> Bool

inRangeOf :: Point3 -> Circle -> Bool

findMostOverlaps :: NESet Circle -> (Int, Point3)

maxOverlapInRegion :: NonEmpty Circle -> V2 Point3 -> (Int, Point3)

inRegion :: Point3 -> V2 Point3 -> Bool

bbSize :: V2 Point3 -> Int

octants :: V2 Point3 -> [V2 Point3]

completeOverlap :: Foldable f => f Circle -> V2 Point3 -> Bool

boundingCube :: V2 Point3 -> [Point3]

day23b :: _ :~> _

circleBounds :: Circle -> [Point3]

boundingBox3 :: Foldable1 f => f Point3 -> V2 Point3

parse23 :: String -> [Circle]