AOC.Challenge.Day24

Description

Day 24. See AOC.Solver for the types used in this module!

After completing the challenge, it is recommended to:

Replace AOC.Prelude imports to specific modules (with explicit imports) for readability.

Remove the -Wno-unused-imports and -Wno-unused-top-binds pragmas.

Replace the partial type signatures underscores in the solution types _ :~> _ with the actual types of inputs and outputs of the solution. You can delete the type signatures completely and GHC will recommend what should go in place of the underscores.

Documentation

data Resist

Constructors

RImmune

RWeak

type Resistance = Map String Resist

data Team

Constructors

TImm

TInf

data Grp

Constructors

G

Fields

_gHP :: Int

_gResist :: Resistance

_gAtk :: Int

_gAtkType :: String

_gInitiative :: Down Int

_gTeam :: Team

type Arena = Map Grp Int

effPower :: Grp -> Int -> Int

stab :: Grp -> Grp -> Int

getDown :: Down a -> a

selectTargets

Arguments

:: Arena

-> Map Grp Grp
targets

makeAttacks :: Map Grp Grp -> Arena -> Arena

fightBattle :: Arena -> Either Arena (Team, Map Grp Int)

day24a :: _ :~> _

day24b :: _ :~> _

type Parser_ = Parsec Void String

parse24 :: Parser_ Arena

teamParser :: Team -> Parser_ Arena

groupParser :: Team -> Parser_ (Grp, Int)

resistanceParser :: Parser_ Resistance