AOC.Common

Description

Common functions for solutions

Documentation

module AOC.Util

iterateMaybe :: (a -> Maybe a) -> a -> [a]

Iterate until a Nothing is produced

(!!!) :: [a] -> Int -> a

Strict (!!)

dup :: a -> (a, a)

A tuple of the same item twice

scanlT :: Traversable t => (b -> a -> b) -> b -> t a -> t b

scanl generalized to all Traversable.

scanrT :: Traversable t => (a -> b -> b) -> b -> t a -> t b

scanr generalized to all Traversable.

firstRepeated :: Ord a => [a] -> Maybe a

Lazily find the first repeated item.

freqs :: (Foldable f, Ord a) => f a -> Map a Int

Build a frequency map

perturbations :: (a -> [a]) -> [a] -> [[a]]

Collect all possible single-item perturbations from a given perturbing function.

perturbations (\i -> [i - 1, i + 1]) [0,10,100]

[ [-1,10,100]

, [ 1,10,100] , [ 0, 9,100] , [ 0,11,100] , [ 0,10, 99] , [ 0,10,101] ]

clearOut :: (Char -> Bool) -> String -> String

Clear out characters not matching a predicate

maximumVal :: Ord b => Map a b -> Maybe (a, b)

Get the key-value pair corresponding to the maximum value in the map

maximumValBy :: (b -> b -> Ordering) -> Map a b -> Maybe (a, b)

Get the key-value pair corresponding to the maximum value in the map, with a custom comparing function.

'maximumVal' == 'maximumValBy' 'compare'

minimumVal :: Ord b => Map a b -> Maybe (a, b)

Get the key-value pair corresponding to the minimum value in the map

minimumValBy :: (b -> b -> Ordering) -> Map a b -> Maybe (a, b)

Get the key-value pair corresponding to the minimum value in the map, with a custom comparing function.

'minimumVal' == 'minimumValBy' 'compare'

maximumValNE :: Ord b => NEMap a b -> (a, b)

Version of maximumVal for nonempty maps.

maximumValByNE :: (b -> b -> Ordering) -> NEMap a b -> (a, b)

Version of maximumValBy for nonempty maps.

minimumValNE :: Ord b => NEMap a b -> (a, b)

Version of minimumVal for nonempty maps.

minimumValByNE :: (b -> b -> Ordering) -> NEMap a b -> (a, b)

Version of minimumValBy for nonempty maps.

deleteFinite :: KnownNat n => Finite (n + 1) -> Finite (n + 1) -> Maybe (Finite n)

Delete a potential value from a Finite.

foldMapPar :: Monoid b => (a -> b) -> [a] -> b

foldMap, but in parallel.

foldMapPar1 :: Semigroup b => (a -> b) -> NonEmpty a -> b

foldMap1, but in parallel.

meanVar :: Fractional a => Fold a (a, a)

Fold for computing mean and variance