AOC.Discover

Description

Template Haskell for discovering all named challenges in a given directory.

Documentation

mkChallengeMap :: [(Finite 25, (Part, SomeSolution))] -> ChallengeMap

Meant to be called like:

mkChallengeMap $$(solutionList "src/AOC/Challenge")

solutionList :: FilePath -> Q (TExp [(Finite 25, (Part, SomeSolution))])

Template Haskell splice to produce a list of all named solutions in a directory. Expects solutions as function names following the format dayDDp, where DD is a two-digit zero-added day, and p is a lower-case letter corresponding to the part of the challenge.

See mkChallengeMap for a description of usage.

type ChallengeMap = Map (Finite 25) (Map Part SomeSolution)

A map of days to parts to solutions.

data ChallengeSpec

A specification for a specific challenge. Should consist of a day and a lowercase character.

Constructors

CS

Fields

_csDay :: Finite 25

_csPart :: Part

dayToInt :: Finite 25 -> Int

Turn a day represented by a Finite 25 into an integer day (1 - 25).

solSpec :: Name -> ChallengeSpec

Get a ChallengeSpec from a given reified solution (name).

solSpec 'day02a == CS { _csDay = 1, _csPart = a }

charPart :: Char -> Maybe Part

Parse a Literal into a Part