AOC.Run.Load

Description

Loading challenge data and prompts.

Documentation

data ChallengePaths

A record of paths corresponding to a specific challenge.

Constructors

CP

Fields

_cpPrompt :: !FilePath

_cpInput :: !FilePath

_cpAnswer :: !FilePath

_cpTests :: !FilePath

_cpLog :: !FilePath

challengePaths :: Integer -> ChallengeSpec -> ChallengePaths

Generate a ChallengePaths from a specification of a challenge.

data ChallengeData

A record of data (test inputs, answers) corresponding to a specific challenge.

Constructors

CD

Fields

_cdPrompt :: !(Either [String] Text)

_cdInput :: !(Either [String] String)

_cdAnswer :: !(Maybe String)

_cdTests :: ![(String, TestMeta)]

challengeData

Arguments

:: Maybe String
session key

-> Integer
year

-> ChallengeSpec

-> IO ChallengeData

Load data associated with a challenge from a given specification. Will fetch answers online and cache if required (and if giten a session token).

countdownConsole

Arguments

:: MonadIO m

=> Integer
year of challenge

-> Finite 25
day to count down to

-> m a
callback on release

-> m a

Run a countdown on the console.

timeToRelease

Arguments

:: Integer
year

-> Finite 25
day

-> IO NominalDiffTime

Get time until release of a given challenge.

showNominalDiffTime :: NominalDiffTime -> String

Pretty-print a NominalDiffTime

charPart :: Char -> Maybe Part

Parse a Literal into a Part

showAoCError :: AoCError -> [String]

htmlToMarkdown :: Bool -> Text -> Either [String] Text

data TestMeta

Constructors

TM

Fields

_tmAnswer :: Maybe String

_tmData :: Map String Dynamic

Parsers

parseMeta :: Parser TestMeta

parseTests :: Parser [(String, TestMeta)]