AOC2018.API

Description

Haskell bindings for Advent of Code 2018 REST API.

API

data API :: Bool -> Type -> Type where

An API command. An API k a an AoC API request that returns results of type a; if k is 'True, it requires a session key. If k is 'False, it does not.

Constructors

APrompt
Fetch prompts, as Markdown.

Fields

:: Finite 25

Day.

-> API False (Map Char Text)

Map of prompts (as markdown). Part 1 is under 'a', Part 2 is under 'b', etc.

AInput
Fetch input

Fields

:: Finite 25

Day.

-> API True Text

ASubmit
Submit answer.

Fields

:: Finite 25

Day.

-> Char

Part. 'a' for part 1, 'b' for part 2, etc.

-> String

Answer. WARNING: not escaped or length-limited.

-> API True (Text, SubmitRes)

Submission reply (as markdown), and result token

data SubmitRes

The result of a submission.

Constructors

SubCorrect

SubWrong

SubWait

SubInvalid

SubUnknown

runAPI :: SessionKey k -> API k a -> IO (Either [String] a)

Run an API command with a given SessionKey to produce the result or a list of (lines of) errors.

WARNING: Does not escape submission answers or limit their length, for ASubmit.

Session Keys

data SessionKey :: Bool -> Type where

Holds a session key. SessionKey 'True contains a session key for sure, but SessionKey 'False may or may not contain one.

Constructors

HasKey :: String -> SessionKey k

NoKey :: SessionKey False

sessionKey :: API k a -> Maybe String -> Maybe (SessionKey k)

Wrap a Maybe String (a possible session key) into a SessionKey based on the expectations of a given API command. If the API command requires a key, Nothing will be returned if no key is given. If the API command doesn't require a key, Just will always be returned.

sessionKey_ :: Maybe String -> SessionKey False

Convert a Maybe String (a possible session key) into a SessionKey 'False --- a SessionKey that may or may not contain a key.