AOC2018.Challenge

Description

Types to drive the challenge runner and help speed up/clean up solutions.

Documentation

data Challenge where

Abstracting over the type of a challenge solver to help with cleaner solutions.

Consists of a parser, a shower, and a solver. The solver solves a general a -> Maybe b function, and the parser and shower are used to handle the boilerplate of parsing and printing the solution.

Constructors

MkC

Fields

:: { cParse :: String -> Maybe a

parse input into an a

, cSolve :: a -> Maybe b

solve an a input to a b solution

, cShow :: b -> String

print out the b solution in a pretty way

} -> Challenge

withSolver :: (String -> Maybe String) -> Challenge

Construct a Challenge from a String -> Maybe String solver, which might fail. Does no parsing or special printing treatment.

withSolver' :: (String -> String) -> Challenge

Construct a Challenge from just a normal String -> String solver. Does no parsing or special printing treatment.

data ChallengeError

Errors that might happen when running a Challenge on some input.

Constructors

CEParse

CESolve

runChallenge :: Challenge -> String -> Either ChallengeError String

Run a Challenge on some input.

runChallenge' :: Challenge -> String -> Maybe String

Run a Challenge on some input, retuning Maybe

type ChallengeMap = Map (Finite 25) (Map Char Challenge)

A map of days to parts to challenges.

data ChallengeSpec

A specification for a specific challenge. Should consist of a day and a lowercase character.

Constructors

CS

Fields

_csDay :: Finite 25

_csPart :: Char