AOC2018.Challenge.Day01

Description

Day 1! Pretty straightforward Haskell pipeline. Parse using manual pattern matching to strip a '+' prefix, because read can't handle positive signs in front of numbers.

The first one is just a simple sum.

The second one lazily computes a running sum with scanl (+) 0 on the infinitely repeated list of inputs (cycle), and then uses a custom utility function firstRepeated to get the first repeated running sum.

Documentation

day01a :: Challenge

day01b :: Challenge