AOC2018.Discover

Description

Template Haskell for discovering all named challenges in a given directory.

Documentation

mkChallengeMap :: [(Finite 25, (Char, SomeSolution))] -> ChallengeMap

Meant to be called like:

mkChallengeMap $$(solutionList "src/AOC2018/Challenge")

solutionList :: FilePath -> Q (TExp [(Finite 25, (Char, SomeSolution))])

Template Haskell splice to produce a list of all named solutions in a directory. Expects solutions as function names following the format dayDDp, where DD is a two-digit zero-added day, and p is a lower-case letter corresponding to the part of the challenge.

See mkChallengeMap for a description of usage.

type ChallengeMap = Map (Finite 25) (Map Char SomeSolution)

A map of days to parts to solutions.

data ChallengeSpec

A specification for a specific challenge. Should consist of a day and a lowercase character.

Constructors

CS

Fields

_csDay :: Finite 25

_csPart :: Char