AOC2018.Interactive

Description

Versions of loaders and runners meant to be used in GHCI.

Documentation

execSolution :: ChallengeSpec -> IO ()

Run the solution indicated by the challenge spec on the official puzzle input.

execSolutionWith

Arguments

:: ChallengeSpec

-> String
custom puzzle input

-> IO ()

Run the solution indicated by the challenge spec on a custom input.

testSolution :: ChallengeSpec -> IO ()

Run test suite for a given challenge spec.

viewPrompt :: ChallengeSpec -> IO ()

View the prompt for a given challenge spec.