AOC2018.Load

Description

Loading challenge data and prompts.

Documentation

data ChallengePaths

A record of paths corresponding to a specific challenge.

Constructors

CP

Fields

_cpPromptUrl :: !FilePath

_cpDataUrl :: !FilePath

_cpPrompt :: !FilePath

_cpInput :: !FilePath

_cpAnswer :: !FilePath

_cpTests :: !FilePath

challengePaths :: ChallengeSpec -> ChallengePaths

Generate a ChallengePaths from a specification of a challenge.

data ChallengeData

A record of data (test inputs, answers) corresponding to a specific challenge.

Constructors

CD

Fields

_cdPrompt :: !(Either [String] String)

_cdInput :: !(Either [String] String)

_cdAnswer :: !(Maybe String)

_cdTests :: ![(String, Maybe String)]

challengeData :: Maybe String -> ChallengeSpec -> IO ChallengeData

Load data associated with a challenge from a given specification. Will fetch answers online and cache if required (and if giten a session token).