AOC2018.Util

Description

Assorted utility functions and orphans used for solutions.

Documentation

strip :: String -> String

Strip trailing and leading whitespace.

iterateMaybe :: (a -> Maybe a) -> a -> [a]

Iterate until a Nothing is produced

(!!!) :: [a] -> Int -> a

Strict (!!)

dup :: a -> (a, a)

A tuple of the same item twice

scanlT :: Traversable t => (b -> a -> b) -> b -> t a -> t b

scanl generalized to all Traversable.

scanrT :: Traversable t => (a -> b -> b) -> b -> t a -> t b

scanr generalized to all Traversable.

eitherToMaybe :: Alternative m => Either e a -> m a

Convert an Either into a Maybe, or any Alternative instance, forgetting the error value.

maybeToEither :: MonadError e m => e -> Maybe a -> m a

Convert a Maybe into an Either, or any MonadError instance, by providing an error value in case Nothing was given.

firstRepeated :: Ord a => [a] -> Maybe a

Lazily find the first repeated item.

freqs :: Ord a => [a] -> Map a Int

Build a frequency map

perturbations :: Ord a => (a -> [a]) -> [a] -> [[a]]

Collect all possible single-item perturbations from a given perturbing function.

perturbations (\i -> [i - 1, i + 1]) [0,10,100]

[ [-1,10,100]

, [ 1,10,100] , [ 0, 9,100] , [ 0,11,100] , [ 0,10, 99] , [ 0,10,101] ]