AOC.Challenge.Day15

Description

Day 15. See AOC.Solver for the types used in this module!

After completing the challenge, it is recommended to:

Replace AOC.Prelude imports to specific modules (with explicit imports) for readability.

Remove the -Wno-unused-imports and -Wno-unused-top-binds pragmas.

Replace the partial type signatures underscores in the solution types _ :~> _ with the actual types of inputs and outputs of the solution. You can delete the type signatures completely and GHC will recommend what should go in place of the underscores.

Documentation

day15a :: _ :~> _

day15b :: _ :~> _