AOC.Challenge.Day17

Description

Day 17. See AOC.Solver for the types used in this module!

After completing the challenge, it is recommended to:

Replace AOC.Prelude imports to specific modules (with explicit imports) for readability.

Remove the -Wno-unused-imports and -Wno-unused-top-binds pragmas.

Replace the partial type signatures underscores in the solution types _ :~> _ with the actual types of inputs and outputs of the solution. You can delete the type signatures completely and GHC will recommend what should go in place of the underscores.

Documentation

type Point = V2 Int

inBounds :: V2 Point -> Point -> Bool

drainMap

Arguments

:: Set Point
clay

-> Int
max y

-> Point
starting point

-> Map Point Bool
map of water, and whether it is draining

fillWater :: Set Point -> Set Point

day17a :: Set Point :~> _

drainWater :: Set Point -> Set Point

day17b :: Set Point :~> _

parseVein :: String -> Set Point

boundingBox :: [Point] -> V2 Point

displayClay :: Set Point -> Set Point -> String